implemented basic arrow drawing on edges
This needs some clean-up for sure, and the arrow head drawing logic
should probably move to inline_primitive. Also, it would probably be
better to just grab the bounding box on the widget that the vertex uses
as a label, then compute the intersection with the edge and only draw
the arrow up to that.